+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Renamed plugin from `rsconnect` to `rsconnect_jupyter`
 - Publish notebooks with source allowing them to be rendered on RStudio Connect.
   `pip` and `virtualenv` are used to determine dependent packages.
+- Deploy logs are shown while notebook is deployed.
 
 # v1.0.1
 - Ensure json is decoded as utf-8